chore: influxdb_client/client/write: fix data_frame_to_list_of_points (#183)
Fix the possibility of data corruption by using a much simpler
regular expression to fix up the results.

Also avoid mutating the DataFrame that's been passed in
by making a shallow copy. This changes semantics
so that if a column mentioned in `PointSettings` exists
in the actual data too, it won't be overridden.

Also change the `test_write_data_frame` test, which is essentially benchmarking
the `data_frame_to_list_of_points` function, so that it just does that
so it can easily be run on a local machine and is independent
of network speed. It runs in about 10s, which is comparable
to the previous performance.