chore: add makefile to repo
This adds a simple makefile to be able to quickly install dependencies,
the library locally as editable for quick development, lint the
necessary files, and run tests with coverage.

It also removes the tox file in favor of a dedicated flake8 file to not
give the impression we are using tox in anyway.