feat: adds the ability to query with SQL
This creates a new SQL Query API that allows for the use of SQL queries
along with InfluxDB that has SQL enabled. Without further testing and
usage it should be expected that this could break easily.

The current model will take the user's URL so they do not need to
provide two different URLs and create two different clients and will
parse out the hostname. This is for use with the gRPC client.

There are no tests as this largely wraps the FlightSQLClient library.